Do not extend a Test class
